Remove hard coded library in CL; configure for RDi (#6)
* Initial commit for a branch to try out RDi and eGit.
RDi has the iProjects perspective which allows testing and development,
and as a side effect, keeps a copy of the source in a stream file.  It
is that stream file we're going to piggyback on.

* Initial commit; using RDi iProjects

* Remove hard coded library

* Put library parameter in CL call

* Change build.txt to match RDi folder hierarchy.
Also, remove PDIR() when done.

* Typo in cleanup command.

* Add MONMSG to handle errors.  
Will ignore if message file already exists.
Will ignore if message ID already exists.